Dei uma arrumada em uns bagui ae da uma olhada ve se ta funcionando

+++ demo/src/main/java/com/example/demo/mapper/Utils/PedidoMapperHelper.java
@@ -62,7 +62,7 @@ public class PedidoMapperHelper {
                     ItemDeCardapio item = itemDeCardapioRepository.findById(dto.getItemId())
                             .orElseThrow(() -> new EntityNotFoundException("Item de cardápio não encontrado."));
 
-                    return item.getPreco().multiply(BigDecimal.valueOf(dto.getQuantidade()));
+                    return (BigDecimal) item.getPreco().multiply(BigDecimal.valueOf(dto.getQuantidade()));
                 })
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
     }
